refactor: Process player in players skills view
