+++ src/modules/playersSkills/services/playersSkillsView.service.ts
@@ -2,35 +2,25 @@ import log from 'loglevel';
 import { Subscription } from 'rxjs';
 import { singleton } from 'tsyringe';
 
-import { FullTrainings } from '@src/common/model/fullTranings.model';
-import {
-    FutureSkillsSummaryWithBest,
-    Player,
-    playerPositionValues,
-    PlayerWithSkillsSummaries,
-    SkillsSummaryCombo,
-} from '@src/common/model/player.model';
+import { Player, PlayerWithSkillsSummaries } from '@src/common/model/player.model';
 import { getPotentialConfig } from '@src/common/model/potential.model';
 import { HtmlElementExtractor } from '@src/common/services/extractors/htmlElement.mapper';
 import { getPlayerMapper } from '@src/common/services/extractors/htmlPlayer.mapper';
-import { PlayersRankService, PlayersRankServiceFactory } from '@src/common/services/playersRank.service';
 import {
     _futurePredicationSettingsKey,
     FuturePredicationSettings,
 } from '@src/common/services/settings/futurePredication.settings';
-import { _rankingsSettingsKey, BaseSettings, RankingsSettings } from '@src/common/services/settings/settings';
+import { _rankingsSettingsKey, BaseSettings } from '@src/common/services/settings/settings';
 import { SettingsChange, SettingsRepository } from '@src/common/services/settings/settings.repository';
 import { DataRow, Header, Table } from '@src/common/services/table.wrapper';
-
-import { FutureSkillsService } from '../../../common/services/futureSkills.service';
-import { SkillCalculatorService } from '../../../common/services/skillCalculator.service';
 
 import { PlayersSkillsFactory } from './playersSkills.factory';
 import { PlayersSkillsTableService } from './playersSkillsTable.service';
 import { PlayersSkillsViewServiceSettings } from './playersSkillsViewServiceSettings';
 import { PageSettingsService, PlayersSkillsViewSettings } from './playersSkillsViewSettings';
+import { ProcessPlayerService } from './ProcessPlayerService';
 
-const _potentialConfig = getPotentialConfig();
+export const _potentialConfig = getPotentialConfig();
 
 /**
  * Przetwarza tabele graczy <Player> z umiejątnościami
@@ -41,12 +31,6 @@ const _potentialConfig = getPotentialConfig();
 export class PlayersSkillsViewService {
     private playersTable?: Table<PlayerWithSkillsSummaries>;
 
-    private fullTrainings?: FullTrainings;
-    private futureAge?: number;
-    private futurePredicationSettings?: FuturePredicationSettings;
-
-    private rankService?: PlayersRankService;
-
     private onSettingsChangedSubs?: Subscription;
 
     private readonly tableService: PlayersSkillsTableService;
@@ -55,11 +39,9 @@ export class PlayersSkillsViewService {
     private settings: PlayersSkillsViewSettings = { hiddenColumns: [] };
 
     constructor(
-        private readonly skillCaluclatorService: SkillCalculatorService,
-        private readonly futureSkillsService: FutureSkillsService,
         private readonly settingsRepository: SettingsRepository,
-        private readonly playersRankServiceFactory: PlayersRankServiceFactory,
         playersSkillsFactory: PlayersSkillsFactory,
+        private readonly processPlayerService: ProcessPlayerService,
     ) {
         log.trace('PlayersSkillsViewService.ctor +');
 
@@ -87,8 +69,8 @@ export class PlayersSkillsViewService {
 
         for (const row of playersTable.rows) {
             const player = row.data;
-            row.data = this.processPlayer(player);
-            this.updatePlayerRanking(row.data.skillsSummaries);
+            row.data = this.processPlayerService.processPlayer(player);
+            this.processPlayerService.updatePlayerRanking(row.data.skillsSummaries);
             row.applyHtmlCells(this.tableService.createSkillsSummaryCells(row.data.skillsSummaries));
         }
 
@@ -103,7 +85,7 @@ export class PlayersSkillsViewService {
         }
 
         for (const row of playersTable.rows) {
-            this.updatePlayerRanking(row.data.skillsSummaries);
+            this.processPlayerService.updatePlayerRanking(row.data.skillsSummaries);
             row.applyHtmlCells(this.tableService.createSkillsSummaryCells(row.data.skillsSummaries));
         }
     }
@@ -140,7 +122,9 @@ export class PlayersSkillsViewService {
             switch (key) {
                 case _futurePredicationSettingsKey:
                     {
-                        this.refreshFuturePredicationSettings(changes[_futurePredicationSettingsKey].newValue);
+                        this.processPlayerService.refreshFuturePredicationSettings(
+                            changes[_futurePredicationSettingsKey].newValue,
+                        );
                         if (this.playersTable) {
                             this.updatePlayers(this.playersTable);
                         }
@@ -149,7 +133,9 @@ export class PlayersSkillsViewService {
                 case _rankingsSettingsKey:
                     {
                         const rankingsSettings = changes[_rankingsSettingsKey].newValue;
-                        this.prepareRankings(rankingsSettings).then(() => this.updateRanking(this.playersTable));
+                        this.processPlayerService
+                            .prepareRankings(rankingsSettings)
+                            .then(() => this.updateRanking(this.playersTable));
                     }
                     break;
                 case this.pageSettingsService.key:
@@ -169,26 +155,10 @@ export class PlayersSkillsViewService {
 
         log.debug('playersSkillsView.service', 'settings', loadedSettings);
         this.settings = { ...this.settings, ...loadedSettings[this.pageSettingsService.key] };
-        this.refreshFuturePredicationSettings(loadedSettings.futurePredicationSettings as FuturePredicationSettings);
-        await this.prepareRankings(loadedSettings.rankingsSettings);
-    }
-
-    private async prepareRankings(rankingsSettings?: RankingsSettings) {
-        if (!rankingsSettings) return;
-        if (!this.rankService) {
-            this.rankService = await this.playersRankServiceFactory.getRanking(rankingsSettings);
-        } else {
-            this.rankService.init(rankingsSettings);
-        }
-    }
-
-    private refreshFuturePredicationSettings(futurePredicationSettings?: FuturePredicationSettings): void {
-        log.trace('refreshSettings:', futurePredicationSettings);
-        if (futurePredicationSettings) {
-            this.futureAge = futurePredicationSettings.futureAge;
-            this.fullTrainings = new FullTrainings(futurePredicationSettings);
-            this.futurePredicationSettings = futurePredicationSettings;
-        }
+        this.processPlayerService.refreshFuturePredicationSettings(
+            loadedSettings.futurePredicationSettings as FuturePredicationSettings,
+        );
+        await this.processPlayerService.prepareRankings(loadedSettings.rankingsSettings);
     }
 
     // #endregion
@@ -231,78 +201,4 @@ export class PlayersSkillsViewService {
         return playersTable;
     }
     // #endregion
-
-    private updatePlayerRanking(skillsSummaryCombo?: SkillsSummaryCombo) {
-        if (!skillsSummaryCombo || !this.rankService) return;
-        if (skillsSummaryCombo.current) {
-            this.updateRankingPerPositon(skillsSummaryCombo.current, this.rankService);
-        }
-        if (skillsSummaryCombo.future) {
-            this.updateRankingPerPositon(
-                skillsSummaryCombo.future,
-                this.rankService,
-                this.futurePredicationSettings?.futureAge,
-            );
-        }
-        if (skillsSummaryCombo.futureMax) {
-            this.updateRankingPerPositon(
-                skillsSummaryCombo.futureMax,
-                this.rankService,
-                this.futurePredicationSettings?.futureAge,
-            );
-        }
-    }
-
-    private updateRankingPerPositon(
-        skillsSummary: FutureSkillsSummaryWithBest,
-        rankService: PlayersRankService,
-        futureAge?: number,
-    ) {
-        playerPositionValues.forEach((positionType) => {
-            const summary = skillsSummary[positionType];
-            summary.rank = rankService.getRank(summary.gs, futureAge);
-        });
-    }
-
-    // Oblicza dodatkowe dane graczy
-    private processPlayer(player: Player): PlayerWithSkillsSummaries {
-        const skillsSummaries: SkillsSummaryCombo = {
-            current: this.skillCaluclatorService.calculateGs(player),
-        };
-        if (skillsSummaries.current && this.futureAge && this.fullTrainings) {
-            if (player.maxTraining) {
-                skillsSummaries.future = this.futureSkillsService.countSkillsInFuture(
-                    skillsSummaries.current,
-                    player.age,
-                    this.futureAge,
-                    this.fullTrainings,
-                    player.maxTraining,
-                    this.futurePredicationSettings,
-                );
-            } else {
-                const potentialDelta = _potentialConfig[player.potential];
-
-                if (potentialDelta) {
-                    skillsSummaries.future = this.futureSkillsService.countSkillsInFuture(
-                        skillsSummaries.current,
-                        player.age,
-                        this.futureAge,
-                        this.fullTrainings,
-                        potentialDelta.min,
-                        this.futurePredicationSettings,
-                    );
-
-                    skillsSummaries.futureMax = this.futureSkillsService.countSkillsInFuture(
-                        skillsSummaries.current,
-                        player.age,
-                        this.futureAge,
-                        this.fullTrainings,
-                        potentialDelta.max,
-                        this.futurePredicationSettings,
-                    );
-                }
-            }
-        }
-        return { ...player, skillsSummaries };
-    }
 }
